requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
colorama==0.4.4
discord.py==1.7.3
discord.py-message-components==1.7.5.4
idna==3.3
lru-dict==1.1.7
multidict==6.0.2
numpy==1.22.3
parameterized==0.8.1
typing_extensions==4.1.1
yarl==1.7.2
